// src/app/tiendas/models/tienda.model.ts
export class TiendaModel {
  idTienda: number;
  nombre: string;
  descripcion: string;
  direccion: string;
  telefono1: string;
  telefono2: string;
  correo: string;
  latitud: number;
  longitud: number;
  regDate: string;
  idUsuario: number;
  ciudad: string;
  departamento: string;
  imagen: string;
}